Fix : 예외처리 하드코딩 제거 / 키값사용

+++ src/main/java/com/example/pstagram/service/friend/FriendService.java
@@ -7,15 +7,18 @@ import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import com.example.pstagram.common.ResponseCode;
 import com.example.pstagram.domain.friend.Friend;
+import com.example.pstagram.domain.friend.FriendStatus;
 import com.example.pstagram.domain.user.User;
-import com.example.pstagram.dto.dto.friend.FriendListResponseDto;
-import com.example.pstagram.dto.dto.friend.FriendResponseDto;
-import com.example.pstagram.dto.dto.friend.FriendWaitingResponseDto;
+import com.example.pstagram.dto.friend.FriendListResponseDto;
+import com.example.pstagram.dto.friend.FriendResponseDto;
+import com.example.pstagram.dto.friend.FriendWaitingResponseDto;
 import com.example.pstagram.exception.friend.DuplicateFriendRequestException;
 import com.example.pstagram.exception.friend.FriendNotFoundException;
 import com.example.pstagram.exception.friend.FriendRequestNotFoundException;
 import com.example.pstagram.exception.friend.SelfRequestException;
+import com.example.pstagram.exception.friend.UnauthorizedException;
 import com.example.pstagram.repository.friend.FriendRepository;
 import com.example.pstagram.service.user.UserService;
 
@@ -27,52 +30,66 @@ public class FriendService {
 	private final UserService userService;
 
 	@Transactional
-	public void requestFriend(Long requesterId, Long recieverId) {
+	public void requestFriend(Long requesterId, Long receiverId) {
 		User requester = userService.getUser(requesterId);
-		User reciever = userService.getUser(recieverId);
+		User receiver = userService.getUser(receiverId);
 
-		if (friendRepository.findByRequesterAndReciever(requester, reciever).isPresent()) {
-			throw new DuplicateFriendRequestException("이미 보낸 요청입니다.");
+		if (friendRepository.findByRequesterAndReceiver(requester, receiver).isPresent()) {
+			throw new DuplicateFriendRequestException(ResponseCode.FRIEND_ALREADY_REQUEST);
 		}// 중복 방지
 
-		if (requesterId.equals(recieverId)) {
-			throw new SelfRequestException("자기 자신에게 친구 요청을 보낼 수 없습니다.");
+		if (requesterId.equals(receiverId)) {
+			throw new SelfRequestException(ResponseCode.SELF_REQUEST);
 		}
-
-		Friend friend = Friend.builder().requester(requester).receiver(reciever).build();
+		if (requesterId == null) {
+			throw new UnauthorizedException(ResponseCode.UNAUTHORIZED);
+		}
+		Friend friend = Friend.builder().requester(requester).receiver(receiver).build();
 
 		friendRepository.save(friend);
 	}
 
 	@Transactional
-	public void cancelRequest(Long requesterId, Long recieverId) {
+	public void cancelRequest(Long requesterId, Long receiverId) {
 		User requester = userService.getUser(requesterId);
-		User reciever = userService.getUser(recieverId);
+		User receiver = userService.getUser(receiverId);
 
-		Friend friend = friendRepository.findByRequesterAndReciever(requester, reciever)
-			.orElseThrow(() -> new FriendRequestNotFoundException("요청이 존재하지 않습니다."));
+		if (requesterId == null) {
+			throw new UnauthorizedException(ResponseCode.UNAUTHORIZED);
+		}
+
+		Friend friend = friendRepository.findByRequesterAndReceiver(requester, receiver)
+			.orElseThrow(() -> new FriendRequestNotFoundException(ResponseCode.FRIEND_NOT_FOUND));
 
 		friendRepository.delete(friend);
 	}
 
 	@Transactional
-	public FriendResponseDto acceptFriend(Long requesterId, Long recieverId) {
-		User reciever = userService.getUser(recieverId);
+	public FriendResponseDto acceptFriend(Long requesterId, Long receiverId) {
+		User receiver = userService.getUser(receiverId);
 		User requester = userService.getUser(requesterId);
 
-		Friend friend = friendRepository.findByRequesterAndReciever(requester, reciever)
-			.orElseThrow(() -> new FriendNotFoundException("요청이 존재하지 않습니다."));
+		if (receiverId == null) {
+			throw new UnauthorizedException(ResponseCode.UNAUTHORIZED);
+		}
+
+		Friend friend = friendRepository.findByRequesterAndReceiver(requester, receiver)
+			.orElseThrow(() -> new FriendNotFoundException(ResponseCode.FRIEND_NOT_FOUND));
 		friend.accept();
 		return new FriendResponseDto(friend.getStatus());
 	}
 
 	@Transactional
-	public FriendResponseDto rejectFriend(Long requesterId, Long currentUserId) {
+	public FriendResponseDto rejectFriend(Long requesterId, Long receiverId) {
 		User requester = userService.getUser(requesterId);
-		User reciever = userService.getUser(currentUserId);
+		User receiver = userService.getUser(receiverId);
 
-		Friend friend = friendRepository.findByRequesterAndReciever(requester, reciever)
-			.orElseThrow(() -> new FriendNotFoundException("요청이 존재하지 않습니다."));
+		if (receiverId == null) {
+			throw new UnauthorizedException(ResponseCode.UNAUTHORIZED);
+		}
+
+		Friend friend = friendRepository.findByRequesterAndReceiver(requester, receiver)
+			.orElseThrow(() -> new FriendNotFoundException(ResponseCode.FRIEND_NOT_FOUND));
 		friend.reject();
 		return new FriendResponseDto(friend.getStatus());
 
@@ -82,7 +99,6 @@ public class FriendService {
 	public List<FriendListResponseDto> getFriendList(Long currentUserId) {
 		User currentUser = userService.getUser(currentUserId);
 
-		//친구가 한명도 없을때 ! , 로그인한 사용자가 아닐때
 		return friendRepository.findFriendList(currentUser.getId())
 			.stream()
 			.map(friend -> FriendListResponseDto.fromFriend(friend, currentUserId))
@@ -93,17 +109,24 @@ public class FriendService {
 	public List<FriendWaitingResponseDto> getWaitingList(Long currentUserId) {
 		User currentUser = userService.getUser(currentUserId);
 
-		//로그인한 사용자가 아닐때, 친구요청이 1명도 없을때
-		return friendRepository.findAllByReceiverAndStatus_Waiting(currentUser).stream()
+		if (currentUserId == null) {
+			throw new UnauthorizedException(ResponseCode.UNAUTHORIZED);
+		}
+
+		return friendRepository.findAllByReceiverAndStatus(currentUser, FriendStatus.WAITING).stream()
 			.map(FriendWaitingResponseDto::fromFriend)
 			.collect(Collectors.toList());
 	}
 
 	@Transactional
-	public void deleteFriend(Long friendId, Long currentUserId) {
+	public void deleteFriend(Long currentUserId, Long friendId) {
+
+		if (currentUserId == null) {
+			throw new UnauthorizedException(ResponseCode.UNAUTHORIZED);
+		}
+
 		friendRepository.deleteById(friendId);
 
-		//로그인한 사용자가 아닐떄
 	}
 }
